refactor: GatheringReceipt jOOQ로 업데이트 처리
