+++ src/main/kotlin/com/server/dpmcore/gathering/gatheringReceipt/application/GatheringReceiptCommandService.kt
@@ -2,8 +2,10 @@ package com.server.dpmcore.gathering.gatheringReceipt.application
 
 import com.server.dpmcore.gathering.gathering.domain.model.Gathering
 import com.server.dpmcore.gathering.gathering.domain.port.inbound.command.ReceiptCommand
+import com.server.dpmcore.gathering.gatheringReceipt.application.exception.GatheringReceiptIdRequiredException
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.GatheringReceipt
-import com.server.dpmcore.gathering.gatheringReceipt.domain.port.GatheringReceiptPersistencePort
+import com.server.dpmcore.gathering.gatheringReceipt.domain.port.inbound.command.GatheringReceiptSplitAmountCommand
+import com.server.dpmcore.gathering.gatheringReceipt.domain.port.outbound.GatheringReceiptPersistencePort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -19,7 +21,10 @@ class GatheringReceiptCommandService(
     ) = gatheringReceiptPersistencePort.save(GatheringReceipt.create(receipt, gathering.id!!), gathering)
 
     fun updateSplitAmount(receipt: GatheringReceipt): Boolean =
-        gatheringReceiptPersistencePort.updateSplitAmount(
-            receipt,
+        gatheringReceiptPersistencePort.updateSplitAmountById(
+            GatheringReceiptSplitAmountCommand.of(
+                receipt.id ?: throw GatheringReceiptIdRequiredException(),
+                receipt.splitAmount,
+            ),
         ) != 0
 }
