+++ src/main/kotlin/com/server/dpmcore/gathering/gatheringReceipt/application/GatheringReceiptQueryService.kt
@@ -3,8 +3,8 @@ package com.server.dpmcore.gathering.gatheringReceipt.application
 import com.server.dpmcore.gathering.gathering.domain.model.GatheringId
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.GatheringReceipt
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.GatheringReceiptId
-import com.server.dpmcore.gathering.gatheringReceipt.domain.port.GatheringReceiptPersistencePort
-import com.server.dpmcore.gathering.gatheringReceipt.domain.port.GatheringReceiptQueryUseCase
+import com.server.dpmcore.gathering.gatheringReceipt.domain.port.inbound.GatheringReceiptQueryUseCase
+import com.server.dpmcore.gathering.gatheringReceipt.domain.port.outbound.GatheringReceiptPersistencePort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
