+++ src/main/kotlin/com/server/dpmcore/gathering/gatheringReceipt/domain/port/inbound/GatheringReceiptQueryUseCase.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.gathering.gatheringReceipt.domain.port
+package com.server.dpmcore.gathering.gatheringReceipt.domain.port.inbound
 
 import com.server.dpmcore.gathering.gathering.domain.model.GatheringId
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.GatheringReceipt
