+++ src/main/kotlin/com/server/dpmcore/gathering/gatheringReceipt/domain/port/outbound/GatheringReceiptPersistencePort.kt
@@ -1,9 +1,10 @@
-package com.server.dpmcore.gathering.gatheringReceipt.domain.port
+package com.server.dpmcore.gathering.gatheringReceipt.domain.port.outbound
 
 import com.server.dpmcore.gathering.gathering.domain.model.Gathering
 import com.server.dpmcore.gathering.gathering.domain.model.GatheringId
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.GatheringReceipt
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.GatheringReceiptId
+import com.server.dpmcore.gathering.gatheringReceipt.domain.port.inbound.command.GatheringReceiptSplitAmountCommand
 import com.server.dpmcore.gathering.gatheringReceipt.infrastructure.entity.GatheringReceiptEntity
 
 interface GatheringReceiptPersistencePort {
@@ -16,7 +17,7 @@ interface GatheringReceiptPersistencePort {
 
     fun findByGathering(gatheringId: GatheringId): GatheringReceiptEntity
 
-    fun updateSplitAmount(gatheringReceipt: GatheringReceipt): Int
+    fun updateSplitAmountById(command: GatheringReceiptSplitAmountCommand): Int
 
     fun findSplitAmountByGatheringId(gatheringId: GatheringId): Int?
 }
